major refactoring to remove a lot of the UI handlers from MainActivity
